requirements.txt
flask==3.0.3
flask-cors==4.0.1
pdfplumber==0.11.4
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
openpyxl==3.1.5
gunicorn==21.2.0
google-generativeai==0.8.3